requirements.txt
annotated-types==0.6.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.3
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
questionary==2.0.1
requests==2.31.0
rsa==4.9
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
